Break the health check logic for the LocalAgentLauncherImpl into its own class
